Nanpa
(Nahm-pah) Nanpa is flirting with someone, or approaching a girl on the
street for a date. You’ll see young yankees (bad boys) at stations
approaching one girl after another in an attempt to get her cell
number. Some of them are looking for hostesses to work in clubs and
cafés, some are simply trying to get a date.